Get airport challenge set up in IntelliJ
I was having trouble getting the folder structure of my tests and source
code set up so that JUnit would work. Took me a while to work out how to
get it set up in IntelliJ, but looks like it's taking care of all of
that now.

Need to work out how to throw/test exceptions next.