Remove builder function in favour of constructors
I thought it was necessary to use the builder pattern to construct my
Airport with multiple optional parameters (capacity and Weather), but
you can do it using overloaded constructors. The problem I thought I had
was that if I was only passing in one parameter then it wouldn't know
whether it was the capacity or the Weather, but in fact it can
distinguish the two cases because the parameters are different types.